ノードIIDを座標値に変更したり、汎用性を高めてみたりする

+++ Src/GameMap.h
@@ -15,7 +15,7 @@ public:
     float offsetX, float offsetZ, float tileSize, const int* data);
   ~GameMap() = default;
 
-  std::vector<glm::ivec2> FindRoute(const glm::ivec2& start, const glm::ivec2& goal) const;
+  std::vector<glm::vec2> FindRoute(const glm::vec2& start, const glm::vec2& goal) const;
   int GetObjectNo(float x, float z) const;
   std::vector<int> Get3x3(float x, float z) const;
